WIP Api refactor
